refactored with usage of timer object
