+++ src/main/java/Game/Main.java
@@ -2,7 +2,6 @@ package Game;
 
 import org.json.JSONObject;
 
-import java.util.ArrayList;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 
@@ -16,47 +15,30 @@ public class Main {
     //TODO final greenery
     //TODO initial buy
 
-    private static String MARS_URL = "";
+    private static final String MARS_URL = "http://168.119.225.172:8080/api/player?id=f814f3933fc2";
 
     private static JSONObject lastJson;
 
     public static void main(String [ ] args) {
         System.out.println("Starting Chatbot");
-        WhatsAppController.connectToWhatsapp();
-
-        //check config in for new Mars Url
-        Executors.newScheduledThreadPool(1).scheduleAtFixedRate(
-                () -> {
-                    try {
-                        MARS_URL = WhatsAppController.getMarsUrl();
-                    } catch (Exception e) {
-                        e.printStackTrace();
-                        WhatsAppController.close();
-                    }
-                },
-                0,
-                30,
-                TimeUnit.SECONDS
-        );
+        WebController.connectToWhatsapp();
 
         Executors.newScheduledThreadPool(1).scheduleAtFixedRate(
             () -> {
                 System.out.println(MARS_URL);
                 JSONObject currentJson = null;
                 try {
-                    currentJson = MarsController.readMarsJson(MARS_URL);
+                    currentJson = WebController.readMarsJson(MARS_URL);
                     if (lastJson == null || lastJson.isEmpty()) {
                         System.out.println("First call to Mars");
-                        lastJson = currentJson;
+                    } else {
+                        System.out.println("Phase: " + JsonEvaluator.getPhase(currentJson) + ",  lastActivePlayers: " + JsonEvaluator.getActivePlayers(lastJson) + " currentPlayers: " + JsonEvaluator.getActivePlayers(currentJson));
+                        JsonEvaluator.evaluateSendingMessage(lastJson, currentJson);
                     }
-                    final ArrayList<String> currentPlayers = MarsController.getActivePlayers(currentJson);
-                    final ArrayList<String> lastActivePlayers = MarsController.getActivePlayers(lastJson);
-                    System.out.println("Phase: " + MarsController.getPhase(currentJson) + ",  lastActivePlayers: " + lastActivePlayers + " currentPlayers: " + currentPlayers);
-                    ChatMakros.evaluateSendingMessage(lastJson, currentJson);
                     lastJson = currentJson;
                 } catch (Exception e) {
-                    System.out.println(lastJson);
-                    System.out.println(currentJson);
+                    System.out.println("lastJson: " + lastJson);
+                    System.out.println("currentJson" + currentJson);
                     System.out.println(e.getMessage());
                 }
             },
